Skip & Requeue pending, unreadable packets during CLOSING
This lets the disconnect packets to float up to the top of the queue immediately, allowing the connection to close gracefully where it might not have been able to before. This will not break up large packets that were already being sent, though.

This commit also contains a slightly faster revision of the Ping packets so you can more accurately determine the speed of the network.