Send InitPacketDeclaration to the server early
This commit has the client send InitPacketDeclaration to the server without waiting for the server to send one of it's own. Because this is still the first packet that both ends would ever recieve, it can be sent early without issue.

The reason for this is to let other non-SubData protocols know that the client is requesting a SubData connection.